requirements.txt
gradio
fpdf2
PyPDF2
nltk
